creating a animation starting point for the vocal channel
Matching the atmosphere of the reverbed, spacey vocals with a spacey blurred image based animation of me singing the lyrics. This image/animation will then be turned into a particle system/array of ellipses so that it can be mapped to the vocal channel and allow movement to be added.

The next step is to figure out the particle/array system and then I can refilm the video so that the lyrics match better for the whole song.