Adding colour and LED screen effect to animation
Adding a blue tone to the images via photoshop to save processing power and get tones I was after.

Then adding a LED Screen effect on top of the video by overlaying a grid of ellipses. I chose to overlay a for loop instead of rasterising the animation like my previous project as it had better performance but gave the same end result.