Adding in Vocal map and time based elements
Going back to the original greyscale image, and using a coloured overlay with the blendMode function on burn mode instead. This allows me more flexibility, with the ability to change the colour in p5.

I then used the lerpColor function to create a glitchy desaturation effect on the face layer with this overlay. This was done by mapping the vocal channel to a lerpColor function that morphs between grey and blue. This function then colours the blended overlay.

I also started to use time based triggers for timing the elements up with the music. The plan for the vocal layer is to re film it so that all the lyrics are correct for the song. I will then apply all the previous effects on and have this face come in only when the vocals are playing to make it responsive to the song.

I also added in a moire style line generator using noise to modulate the paths of lines. This effect can be modulated to be minimal or complex, and appear fast or slow. This gives me flexibility to reflect the different sections of the song. The large amount of adjustable parameters also make it easy to create variations and modulate with the stem channels.

The next step for the moire patterns is to create additional time based changes and add in modulators.