Adding more time based controls with maps and pauses and adding variations
Adding a time based map to the pattern to match the speed and atmospheric beginning of the song and then have it build up towards the body, matching the feeling of the build up. This was done by mapping the seconds variable I created  to the lineSpace paraeater that controls how close lines in the pattern are to each other. I started it off at 0 seconds with the lines sparsely apart to give more of the atmospheric feel and then slowly decreased the spacing till it reached the body where it was the full pattern. This gave the feeling of building up and the development that the song has.

I also added in pauses that match drum pauses etc in the song and added initial pattern variations.

The next step is to create more variation by adding in a different shape instead of the lines to give variety in the second half of body 1. This way also allows me to use the same body of code, keeping processing power down.

I also want to add in bass, other and drum element maps to parameters of the pattern to give further variations.

I still also need to re film /time the face layer to match the vocals and have it come in and out. Will do this after figuring out the second shape.

I Also need to clean up the code into their own functions to make it easier to follow.