Adding in rectangles do to the pattern generator for more variation
Adding in rectangles to the pattern generator and having them in the 3rd and 4th 16s to give variation to the visualiser and reflect the change in musical elements that come in at the 3rd 16. This was a simple way of modifying the code I already had and it produces a strong contrast to the line version while keeping performance down since its an animated work and that matters.

I still need to add in the bass, drum, other maps to modulate some parameters with them for further variation.

Also still need to match the vocals up with the face layer.