testing how to load the new video that syncs to all the lyrics - qodo code for loading images as object instead of array
I tried loading the frames in like my previous version but it was too laggy as it was trying to load 2000 images all at once.

To fix this I used qodo to help me figure out how to load the images in small batches in real time. It swapped my array setup for an object instead. Apparently objects don't waste memory like an array does and so you can skip frames etc. This fixed it and allowed it to run smoothly.

The only thing qodo added was how to load the images. The rest of the code is still the same as I wrote it before.

The next step is to modify the qodo code so that is more understandable for me with better variables, I may try simplify it as well, before then loading in the other sections of the vocals.