modifying and simplfiying the qodo code for loading the images
Simplifying the qodo code for loading the images where I can and changing the variable names to make more sense to me.

I also cleaned up some code in the moire sections, combining if statements together for sections that remain the same

next step is to load all of the sections of the face animations, now that I have got it working.

Will also need to clean everything up and put it all the code into functions to make it easier to keep track of.

after that, I can add in more maps, and start working on the second half of the song