loading all face layer sections that are timed to the vocal layer
Loading all of the face layers that are tied to the vocal layer and getting it to not crash.  I also added in a fade at the end of the vocal animations so they didn't abruptly end. This made it look better and also reflected the atmospheric feel of the reverbed vocals.

Using the modified loading code from qodo was too processor heavy and would crash.

I firstly swapped the jpgs to webp since the file size of webp could be less than 50kb. This helped alot but the visualiser would still crash towards the end of the song.

I tried using qodo/Ai to fix it, but it would not work, it kept giving results that crashed, just at different points of the song. I eventually asked my mother in law who's a software tester to have a look and she spotted it instantly. The images were being loaded in batches which was great but then they weren't being cleared from the memory after use, meaning that by the end of the song thousands of images were in the memory when they weren't needed, crashing it.

With the help of my mother in law,I managed to code a way for the images to be wiped from memory when not needed anymore. It essentially counts the frames from a 0 point and then once it reaches 120 frames ( 2 seconds), it checks to see if that image is needed anymore and wipes it if its not.This instantly fixed the crashing.

My mother in law suggested I could clean up the code more by using an array of objects for the faceframes, instead of listing out all the face frame objects in a row, this made the code more legible as each object could be displayed as one line horizontally as opposed to 4 or 5 vertically per object. since there is 27 objects, this took that code from over approx 100 lines, down to 27.

she also suggested other ways to clean up the code, such as using the for (let x of y) function. This allowed for things to be called in one line automatically rather than hardcoding everything. This reduced down the amount of code needed to achieve the same thing.  So instead of a 27 if statements, one line of code can be used to do the same thing.

 She also suggested better variable names as she had a hard time understanding what I meant, so I changed those to make it easier for myself to understand the new code I had learnt better. This was helpful as changing the names, allowed me to understand what the code was doing by having to think about each line.

I also cleaned up the code more, eliminating redundant code and merging functions.

The plan is also to clean up the moire pattern if statements using a for (let x of y) function once I have figured out the second half of the visualiser.

Now that the vocal face layer is sorted the next step is to add in the patterns for the second half of the song and then add more maps in from the stem channels.