Added in patterns for second half of the song + some more maps
Adding in the patterns for the second half of the song. Following the same structure as the song which flips the body sections around, eg body 1 has sections 1,2,3,4 and body 2 copies these then flips these to be 3,4,1,2 to give variation. I followed this by using the rectangles for the first half (3,4) and the lines for the second half (1,2). essentially flipping the way the pattern is played in the first half.

 I also changed the rectangles to ellipses to introduce a new surprise element that only comes in in the second half. This added more variation to the whole song, making it appear less repetitive since its a song of over 5 minutes.

I also added in maps to the drums, bass and other channels, mapping them to the noise modulators that control how the pattern is created. I had to map the values subtly with very small numbers to keep it from turning into random noise and retain the flow I liked.

The next step is to adjust more pattern parameters for the second half of the song to give greater variation.

I will also add in some maps to the pattern colours to match the desaturated glitch effect of the vocal face layer.