further tweaking the second half to make varied and adding in more maps
Adding in maps to the colour so that there is a sense of build up in the intro and breakdowns by using if statements to have pink only happen when the drums are over a certain volume threshold. This gave this sense of progression that matched the build up. I did it with if statements instead of maps since I didn't like the in between colours that the lerpColor created and instead just uses if (drums < 40) have it grey. This also added to the glitchy feel when the drums started to come in since they were going up and down in volume a bit causing flickering.

I also mapped the drum volume to the blue pattern so that there is a subtle desaturated glitch effect on the blue layer.

I also tweaked the second half's pattern modulators to give subtle variation form the first half. This made it less repetitive while still remaining cohesive and reflect the songs subtle change between body 1 and body 2.

This subtle change was achieved by changing the angle that the pattern is generated at. changing values being mapped from the drum, bass, other channels. and adjusting the noise Map parameters which also effect the layout of the pattern.

the next step is to fix the predrop vocal face section since it didn't load properly and is static for a bit.

And then to add some subtle grain over the whole sketch to match the patterns to the face layer.